Modifications for libecc.
git-svn-id: http://ellcc.org/svn/ellcc/trunk@2260 82857169-6e38-49d7-aaa8-07fb558f12fc

+++ libecc/include/ppc/linux/sys/types.h
@@ -165,7 +165,7 @@ typedef	long		key_t;		/* IPC key (for Sys V IPC) */
 
 typedef	uint32_t	mode_t;		/* permissions */
 
-typedef	uint32_t	nlink_t;	/* link count */
+typedef	uint16_t	nlink_t;	/* link count */
 
 #ifndef	off_t
 typedef	int32_t	off_t;		        /* file offset */
